Extract Home header controls into shared page-header components
